Refactor into type modules
